backend/requirements/base.txt
boto3==1.16.23

celery==5.0.2
kombu==5.0.2

channels==3.0.2
channels-redis==3.2.0

daphne==3.0.1

graphene-django==2.13.0
django-graphql-jwt==0.3.1

Django==3.1.3
django-filter==2.2.0
djangorestframework==3.12.2
djangorestframework_simplejwt==4.4.0
django-debug-toolbar==2.2
django-model-utils
gevent==20.9.0

# Python Social Auth

social-auth-app-django==3.4.0
social-auth-core==3.3.3

# Django page caching is not working with this version of Redis
# but this version is required by Celery
django-redis==4.12.1

django-storages==1.10.1

# TODO: figure out why pandas is taking a long time to download
# in cdk deploy step
# pandas==0.24.2

gunicorn==20.0.4

psycopg2-binary==2.8.6

redis==3.5.3

requests==2.24.0
